first predict model
Hi guys, here's our first predict model! 
So there're 69 taxi zones in NYC and we'll create every pickel file for each taxi zone, just like what we did in the dublin bike project.
And this is the first one, I will continue to upload the rest to this drive。
Inputs to this model are 3 parameters: temperature, month and time of week:
like time_of_day group by 15 mins，eg:00:00:00AM is 0, 00:15:00 is 1 and 00:30:00 is 2.....
month: from 1-12 stands for Jan to Dec
temperature is the temperature the time user input
output will be calm rate of every zone, which is a number between 0 and 1
For now we have average calm rate of every zone, we also will have more specific information like precise calm rate to each street node. 
We will be working on these in more detail over the next few days。